feat: Implement dynamic product and cart functionality
It adds the ability to fetch product data dynamically,
updates the product description component to use a form, and introduces a new Button component.
It also adds the ability to add items to the cart and creates new tables for carts, products, and users.
The hover effect on the catalogue page has been updated and the number of roducts displayed on the home page is now adjusted based on the screen size.